Added Artwork Loading
The artwork is loaded from a network location using picasso. The BitmapFetcher might be overly complex.

When testing in the emulator, the notification can flicker when spamming play pause button. It seemed to remove the artwork before handling the action. It might just be an emulator quirk. Need to test on real device to figure this out.

Next up: adding items to the queue.