Partially Working PlaybackService
It uses a modified version of ExoPlayer with changes from this PR.

https://github.com/google/ExoPlayer/pull/5328

It seems that Exoplayer decided overloading the notification cancellation action
to also stop playback was a good idea which leads to some confusing
behavior.

I think burning the notification manager to the ground and re-writing it
in this application is a good idea.