Call Stop Self in Service
A service can be bound or started or both.

When a service is started, it will not be killed by the
system. A system is started with startService and friends
(startForegroundService, Utils.startService).

A service can also be bound or unbound. A service is bound
when there is at least one client bound to it with
bindService.

A service can be either foreground or background. A service
is background by default. It can go to the foreground state
by calling startForeground and return to the background
state by calling stopForeground(bool). Background services
can be killed by the system sometimes (more often in O+),
even if started.

A service can be killed by the system only if the service is
both not started and not bound.

START_STICKY tells the system to restart the service later if
it is killed because of low memory. The documentation
for START_STICKY recommends this for media playback apps.